Migrate to interface, structure refactoring
