+++ app/src/main/java/edu/kmaooad/capstone23/communication/services/ParticipantServiceImpl.java
@@ -4,7 +4,7 @@ import edu.kmaooad.capstone23.communication.dal.entities.Participant;
 import edu.kmaooad.capstone23.communication.interfaces.repositories.ParticipantRepository;
 import edu.kmaooad.capstone23.communication.interfaces.services.ChatService;
 import edu.kmaooad.capstone23.communication.interfaces.services.ParticipantService;
-import edu.kmaooad.capstone23.users.services.UserService;
+import edu.kmaooad.capstone23.users.interfaces.services.UserService;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
