+++ app/src/test/java/edu/kmaooad/capstone23/orgs/members/TestWithDbClearance.java
@@ -3,7 +3,7 @@ package edu.kmaooad.capstone23.orgs.members;
 import edu.kmaooad.capstone23.experts.dal.ExpertsRepository;
 import edu.kmaooad.capstone23.members.dal.MembersRepository;
 import edu.kmaooad.capstone23.orgs.dal.OrgsRepository;
-import edu.kmaooad.capstone23.users.interfaces.UserRepository;
+import edu.kmaooad.capstone23.users.interfaces.repositories.UserRepository;
 import jakarta.inject.Inject;
 import org.junit.jupiter.api.AfterEach;
