+++ app/src/test/java/edu/kmaooad/capstone23/orgs/members/TestWithMembersSetUp.java
@@ -4,7 +4,7 @@ import edu.kmaooad.capstone23.members.dal.Member;
 import edu.kmaooad.capstone23.members.dal.MembersRepository;
 import edu.kmaooad.capstone23.orgs.dal.Org;
 import edu.kmaooad.capstone23.orgs.dal.OrgsRepository;
-import edu.kmaooad.capstone23.users.interfaces.UserRepository;
+import edu.kmaooad.capstone23.users.interfaces.repositories.UserRepository;
 import edu.kmaooad.capstone23.users.mocks.UserMocks;
 import jakarta.inject.Inject;
 import org.bson.types.ObjectId;
